+++ store-deno-kv-fs/mod.ts
@@ -1,6 +1,10 @@
 import * as kv from "@jollytoad/store-deno-kv";
 import * as fs from "@jollytoad/store-deno-fs";
-import type { StorageKey, StorageModule } from "@jollytoad/store-common/types";
+import type {
+  MinimalStorageModule,
+  StorageKey,
+  StorageModule,
+} from "@jollytoad/store-common/types";
 
 export type { StorageKey, StorageModule };
 
@@ -14,7 +18,7 @@ export type { StorageKey, StorageModule };
   clearItems,
   close,
   url,
-}) satisfies StorageModule;
+}) satisfies MinimalStorageModule;
 
 /**
  * Returns the `import.meta.url` of the module.