+++ store-web-storage/mod.ts
@@ -1,5 +1,9 @@
 import { fromStrKey, toStrKey } from "@jollytoad/store-common/key-utils";
-import type { StorageKey, StorageModule } from "@jollytoad/store-common/types";
+import type {
+  MinimalStorageModule,
+  StorageKey,
+  StorageModule,
+} from "@jollytoad/store-common/types";
 
 export type { StorageKey, StorageModule };
 
@@ -13,7 +17,7 @@ export type { StorageKey, StorageModule };
   clearItems,
   close,
   url,
-}) satisfies StorageModule;
+}) satisfies MinimalStorageModule;
 
 const SEP = "/";