+++ store-web-storage/store.test.ts
@@ -1,9 +1,11 @@
 import {
   open,
   testClearItems,
+  testCopyItems,
   testGetItem,
   testHasItem,
   testListItems,
+  testMoveItems,
   testOrdering,
   testRemoveItem,
   testSetItem,
@@ -21,6 +23,8 @@ Deno.test("store-web-storage", async (t) => {
     await testListItems(t, store);
     await testRemoveItem(t, store);
     await testClearItems(t, store);
+    await testCopyItems(t, store);
+    await testMoveItems(t, store);
     await testOrdering(t, store);
   } finally {
     await store.close();