+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0]
+
+### Added
+
+- `copyItems` and `moveItems` optional extended storage module functions
+- [store-common/test-storage-module] common tests for these new functions
+- [store-common/copy-items] fallback implementation for `copyItems`
+- [store-common/move-items] fallback implementation for `moveItems`
+- [store-common/types] new types to represent storage modules with/without
+  extended features
+
 ## [0.5.0]
 
 ### Added